third_party/WebKit/LayoutTests/fast/css/invalidation/full-page-media.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
.outer #inner:not(:-webkit-full-page-media) {
    background-color: green
}
#outer + div {
    color: pink
}
</style>
<div id="outer">
    <div>
        <div id="inner"></div>
    </div>
    <div></div>
    <div></div>
</div>
<div></div>
<script>
description("Support style invalidation when selectors contain :-webkit-full-page-media")

var transparent = "rgba(0, 0, 0, 0)";
var green = "rgb(0, 128, 0)";

shouldBe("getComputedStyle(inner).backgroundColor", "transparent");

inner.offsetTop; // force recalc
outer.className = "outer";

if (window.internals)
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");

shouldBe("getComputedStyle(inner).backgroundColor", "green");
</script>
