third_party/WebKit/LayoutTests/fast/css/invalidation/full-page-media-expected.txt

Support style invalidation when selectors contain :-webkit-full-page-media

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(inner).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(inner).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
